Separated free-form query param tests into their own suite
